Add event listener to insert tracking code on each page
Adds a ViewEvent susbcriber which appends a Google Analytics async tracking
script to the <head> element.  Since page templates are children of the layout
ViewModel, we check to see when we are rendering the root layout and append the
script tag there.  There is an additional check for templates containing the
word 'admin' to exclude analytics from those pages.